Remove common libraries.
This removes /common directory.

After the move to newlib, the libc was used only by the kernel. It
has thus been moved to kernel/lib, where the rest of the code has
also been rationalised.

The only part used in userspace was atomic in libmrg/vm.c, which
was unneeded anyway.

The squoze part (common/lib/misc) has been duplicated and moved
to libsquoze for userspace and ukern/lib/string for kernel.